Spring has already validated the token in TokenResource, so revoke as-is

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/tokenverification/service/TokenService.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.tokenverification.service
 
 import jakarta.validation.ValidationException
 import org.slf4j.LoggerFactory
+import org.springframework.security.oauth2.jwt.Jwt
 import org.springframework.security.oauth2.jwt.JwtDecoder
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.tokenverification.data.Token
@@ -45,8 +46,12 @@ class TokenService(private val tokenRepository: TokenRepository, private val jwt
     tokens.forEach { tokenRepository.delete(it) }
   }
 
-  fun revokeToken(jwt: String) {
-    val (jwtId, _) = validateJwt(jwt)
+  fun revokeToken(jwt: Jwt) {
+    val jwtId = jwt.getClaimAsString("jti")
+    if (jwtId.isNullOrBlank()) {
+      log.info("Unable to retrieve jwt id due to jti being null or blank")
+      throw ValidationException("Unable to find jwtId from token")
+    }
     log.info("Revoking token with jwtId of {}", jwtId)
     tokenRepository.deleteById(jwtId)
   }
